Move still-used code from experimental.regrid to analysis._area_weighted.
